+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed example in Set-QualysAssetGroups
+- Fixed type bug in Get-QualysScans
 
 ## [1.2.7] - 2021-04-01