+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.2.8] - 2021-04-05
+
+### Changed
+
+- Fixed example in Set-QualysAssetGroups
+- Fixed type bug in Get-QualysScans
+- Fixed bug with parameter name in Add-QualysScanSchedule
+
 ## [1.2.7] - 2021-04-01
 
 ### Added